+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## v2.15.0 - 2023-07-10
+
 ### Fixed
 
 - Model Registry bug where cache size option was ignored